+++ core/env/prefs.sh
@@ -10,10 +10,12 @@ load_prefs() {
         BANNER_NAME=$(grep "^NAME=" ~/.username | cut -d= -f2- | sed 's/^"//;s/"$//')
         CUR_FONT=$(grep "^FONT=" ~/.username | cut -d= -f2- | sed 's/^"//;s/"$//')
         CUR_CAT_STYLE=$(grep "^CAT=" ~/.username | cut -d= -f2- | sed 's/^"//;s/"$//')
+        CUR_PROMPT_STYLE=$(grep "^STYLE=" ~/.username | cut -d= -f2- | sed 's/^"//;s/"$//')
         CUR_CHANNEL=$(grep "^CHANNEL=" ~/.username | cut -d= -f2- | sed 's/^"//;s/"$//')
         [[ -z "$BANNER_NAME" ]] && BANNER_NAME="Promptify"
         [[ -z "$CUR_FONT" ]] && CUR_FONT="random"
         [[ -z "$CUR_CAT_STYLE" ]] && CUR_CAT_STYLE="full"
+        [[ -z "$CUR_PROMPT_STYLE" ]] && CUR_PROMPT_STYLE="parrot"
         [[ -z "$CUR_CHANNEL" ]] && CUR_CHANNEL=$(resolve_channel)
     fi